initial commit
includes logic to build a custom manifest file that can be imported to components on runtime.

utils/manifest.js will run on build and generate a file in 'assets' called 'manifest.js', which will output a JSON object with basic post info.

this info/object is used to validate dynamic routes, so static routes can be generated, and a static blog can be run using simple markdown files.